+++ examples/provider/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     azexecute = {
       source  = "dyntora/azexecute"
-      version = "~> 0.7"
+      version = "~> 0.8"
     }
   }
 }