+++ examples/resources/azexecute_application/resource.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     azexecute = {
       source  = "dyntora/azexecute"
-      version = "~> 0.7"
+      version = "~> 0.8"
     }
   }
 }
@@ -23,6 +23,15 @@ resource "azexecute_application" "deployment" {
   web_redirect_uris              = ["https://platform.example.com/signin-oidc"]
   web_enable_id_token_issuance   = true
   requested_access_token_version = 2
+  app_roles = [{
+    id                     = "11111111-2222-4333-8444-555555555555"
+    display_name           = "Deployment Reader"
+    value                  = "Deployment.Reader"
+    description            = "Reads deployment status."
+    is_enabled             = true
+    allow_users_and_groups = true
+    allow_applications     = true
+  }]
 
   api_permission_request {
     target_type                      = "ExternalApi"